@@ -1,17 +1,17 @@
 name: flutter_ios_calculator
 description: iOS 10 and iOS 11&#x2F;12 style calculators for Flutter
-version: 0.0.2
+version: 0.1.0
 homepage: https://github.com/ominibyte/flutter_ios_calculator.git
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  auto_size_text: ^2.1.0
-  intl: ^0.16.0
-  flutter_device_type: ^0.1.0
+  auto_size_text: ^3.0.0
+  intl: ^0.17.0
+  flutter_device_type: ^0.4.0
 
 dev_dependencies:
   flutter_test:
